lib: Separate out cpumask helpers that depend on bpf_cpumask_populate
On older kernels bpf_cpumask_populate is not available. Split out the
arena cpumask helpers into a separate file for schedulers that need to
convert arena cpumasks into bpf cpumasks (wd40). This provides better
compatibility for schedulers that only need arena cpumask operations
without the need of converting arena cpumasks into bpf cpumasks.
