+++ scheds/rust/scx_wd40/build.rs
@@ -16,6 +16,7 @@ fn main() {
         .add_source("../../../lib/arena.bpf.c")
         .add_source("../../../lib/bitmap.bpf.c")
         .add_source("../../../lib/cpumask.bpf.c")
+        .add_source("../../../lib/cpumask_bpf.bpf.c")
         .add_source("../../../lib/sdt_task.bpf.c")
         .add_source("../../../lib/sdt_alloc.bpf.c")
         .add_source("../../../lib/topology.bpf.c")
